codegen: extract compute_inherits + rewrite_single_field_objects_to_tuples to shared
These two helpers were going to be duplicated, character-for-character,
between the existing java.rs target and a future csharp.rs target. Pull
them into shared up front so the next language to land doesn't need its
own copy.

Verified that the Java output against the current spec
(turbopuffer-d1a77644…dc613856c.yml) is byte-identical before and
after the refactor.
